refactor: align project structure with FSD and fix hydration issues
