+++ src/app/(auth)/signup/SignupPage.styles.ts
@@ -1,0 +1,11 @@
+import { Box } from "@mui/material";
+import { styled } from "@mui/material/styles";
+
+export const AuthWrapper = styled(Box)({
+  minHeight: '100vh',
+  display: 'flex',
+  alignItems: 'flex-start',
+  justifyContent: 'center',
+  backgroundColor: 'rgba(53, 53, 53, 1)',
+  paddingTop: '0.5rem',
+});
