+++ src/app/(main)/layout.tsx
@@ -1,5 +1,9 @@
-import { AppLayout } from "@/widgets/app-layout/AppLayout";
+import { AppLayout } from "@/widgets/app-layout/ui/AppLayout";
 
-export default function MainLayout({ children }: { children: React.ReactNode }) {
+export default function MainLayout({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
   return <AppLayout>{children}</AppLayout>;
 }
