+++ src/widgets/app-layout/ui/AppLayout.styles.ts
